fix(coding-agent): stabilize adaptive stream pacing
